Use a standard static site generator and overhaul site (#28)
* Migrate content to zola

This currently is only a skeleton, any design is missing.

* actions: Update nix

* Enable atom feeds

* Add iCal feed

It is not very pretty, but it should work.

Every announcement can optionally have event metadata. All those
announcements should also have a plain-text description that will be
shown in the iCal feed.

* Streamline flake

* Use NixOS common styles again

This also switches to less, which sadly is less integrated with zola.

Co-authored-by: Simon Bruder <simon@sbruder.de>

* Use Makefile

This simplifies the flake and more easily allows development outside of
a Nix build.

* Update readme to reflect current development workflow

* Re-import old styling

* Add navbar

* Re-import footer

* Apply reasonable styling to blog/announcements

* Add shortcode for youtube embedding

For some reason it is not possible to override Zola’s internal youtube
shortcode, so this needs to be called differently.

It can be used to either embed videos or full playlists. Videos can be
embedded like this:

{{ yt(id="h8hWX_aGGDc") }}

Playlists can be embedded like this:

{{ yt(list="PLt4-_lkyRrOO9M5dffzlLiSQrHGdbKtHQ") }}

* Add live streaming page

Co-authored-by: nrdsp <nrdsp@users.noreply.github.com>

* License code samples under CC0/public domain

* Add optional author attribution to posts

* Use relative URL as base path

Otherwise branch previews are broken.

* Reduce CLS

* videos: Add first recording of Public Lecture Series

Co-authored-by: Ilan Joselevich <personal@ilanjoselevich.com>
Co-authored-by: nrdsp <nrdsp@users.noreply.github.com>